+++ src/main/java/DukeException.java
@@ -1,0 +1,6 @@
+public class DukeException
+        extends Exception {
+    public DukeException(String errorMessage) {
+        super(errorMessage);
+    }
+}
